Broke apart customer analytics module
